Apollo_Lookup_Bot/requirements.txt
pyTelegramBotAPI
gspread
oauth2client
python-telegram-bot==20.0